% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tex.R
\name{add_section}
\alias{add_section}
\title{Add a section}
\usage{
add_section(x, tlfs)
}
\arguments{
\item{x}{the name of the section to add (either "section", "subsection", or "subsubsection")}

\item{tlfs}{Data frame with columns for section titles for each TLF
("section", "subsection", "subsubsection") and "path" with the path to the
pdf file for the TLF.}
}
\description{
Add a section
}
