% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tex.R
\name{build_cover_toc_tex}
\alias{build_cover_toc_tex}
\title{Build the cover page and TOC .tex}
\usage{
build_cover_toc_tex(
  study,
  open_closed,
  purpose,
  mtg_type,
  mtg_date,
  mtg_times,
  cutoff_snapshot,
  dco_date,
  conf_statement
)
}
\arguments{
\item{input}{A list with any of the following named elements. If an element
is not provided. It will be left blank in the output.
1. \code{study} The name of the study to use as the title
2. \code{open_closed} "Open" or "Closed" for the open or closed reports.
3. \code{purpose} Usually "DMC Meeting"}
}
\description{
Build the cover page and TOC .tex
}
