% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tex.R
\name{build_tex}
\alias{build_tex}
\title{Build LaTeX document of concatenated report}
\usage{
build_tex(path_tlf_csv, toc = TRUE, coverpage = NULL, path_tex)
}
\arguments{
\item{path_tlf_csv}{Path to a CSV file with with two columns: "title" with
the title of each TLF and "path" with the path to the pdf file for the TLF.}

\item{toc}{Should a table of contents be included?}

\item{coverpage}{A character string of LaTeX to include for a coverpage.}
}
\description{
Build LaTeX document of concatenated report
}
