% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{Concatenate PDF via LaTeX}
\usage{
concat(path_tlf_csv, toc = TRUE, coverpage = NULL, path_tex, path_pdf, ...)
}
\arguments{
\item{path_tlf_csv}{Path to a CSV file with with two columns: "title" with
the title of each TLF and "path" with the path to the pdf file for the TLF.}

\item{toc}{Should a table of contents be included?}

\item{coverpage}{A character string of LaTeX to include for a coverpage.
E.g., generated using \code{\link[concat]{add_cover_tex}}}

\item{path_tex}{The path to output the LaTeX document}

\item{path_pdf}{The path to output the PDF document}

\item{...}{Further arguments passed to \code{\link[tinytex]{pdflatex}}}
}
\description{
Concatenate PDF via LaTeX
}
