% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{concat_setup}
\alias{concat_setup}
\title{Install tinytex and required LaTeX packages}
\usage{
concat_setup(pkgs = c("geometry", "pdfpages", "hyperref"))
}
\description{
Install tinytex and required LaTeX packages
}
