% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_times.R
\name{format_date}
\alias{format_date}
\title{Format date}
\usage{
format_date(date_in, format = "\%A \%B \%d, \%Y")
}
\arguments{
\item{date_in}{The date to input in ISO 8601, e.g., "2021-12-13"}

\item{format}{Date format to use, see \code{\link[base]{strptime}}}
}
\description{
Format date
}
