% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_times.R
\name{format_duration}
\alias{format_duration}
\title{Print meeting duration}
\usage{
format_duration(tzone, start_time, end_time)
}
\arguments{
\item{tzone}{The time zone to display the meeting start and end times}

\item{start_time}{Meeting start date & time}

\item{end_time}{Meeting end date & time}
}
\description{
Print meeting duration
}
