% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pps.R
\name{format_pps}
\alias{format_pps}
\title{Format PPS}
\usage{
format_pps(pps, open_closed, path_prod)
}
\arguments{
\item{pps}{Data frame of the unformatted PPS}

\item{open_closed}{String that identifies which report to produced. Usually
"open" or "closed"}

\item{path_prod}{Path to the folder that has the subfolders of the outputs to
produce}
}
\description{
Format PPS
}
