% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_times.R
\name{format_times}
\alias{format_times}
\title{Format times}
\usage{
format_times(meeting_date, start_time, end_time, tzone, tzones)
}
\arguments{
\item{start_time}{String with the start time in the time zone \code{tzone}.
Use 24-hour time. (I.e., "13:00" not "1:00 PM").}

\item{end_time}{String with the start time in the time zone \code{tzone}. Use
24-hour time.}

\item{tzones}{The time zones you wish to display. See
\url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for time
zone names you can use.}
}
\description{
Create a character vector with meeting start and end times in different time
zones.
}
\examples{
format_times("2021-07-02", "7:00", "9:00", "America/Los_Angeles",
 c("America/Los_Angeles", "America/New_York", "Europe/London"))
}
