% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cover_tex.R
\name{gen_cover_tex}
\alias{gen_cover_tex}
\title{Generate cover page LaTeX string}
\usage{
gen_cover_tex(
  study,
  report_title,
  purpose,
  mtg_type,
  mtg_date,
  start_time,
  end_time,
  tzone,
  tzones,
  cutoff_snapshot,
  dco_date,
  conf_statement
)
}
\arguments{
\item{study}{The name of the study, e.g., "Sponsor Protocol"}

\item{report_title}{The title of the report, e.g., "Open Report"}

\item{purpose}{E.g., "DMC Meeting"}

\item{mtg_type}{What kind of meeting is it? E.g., "Teleconference"}

\item{mtg_date}{The date of the meeting in ISO 8601 format, e.g.,
"2021-12-10".}

\item{start_time}{String with the start time in the time zone \code{tzone}.
Use 24-hour time. (I.e., "07:00" not "7:00" and "13:00" not "1:00 PM").}

\item{end_time}{String with the start time in the time zone \code{tzone}. Use
24-hour time.}

\item{tzone}{The time zone that the start and end times are in see the link
in \code{tzones} for names to use.}

\item{tzones}{The time zones you wish to display. See
\url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for time
zone names you can use.}

\item{cutoff_snapshot}{"Cutoff" if cutoff date is used "Snapshot" if snapshot date is used.}

\item{dco_date}{Data cutoff or snapshot date in ISO 8601, e.g., "2021-11-10".}

\item{conf_statement}{Statement of confidentiality.}
}
\description{
Generate cover page LaTeX string
}
