% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cover_tex.R
\name{paste_cover_tex}
\alias{paste_cover_tex}
\title{Generate cover page LaTeX string}
\usage{
paste_cover_tex(
  study,
  report_title,
  purpose,
  mtg_type,
  mtg_date,
  mtg_times,
  cutoff_snapshot,
  dco_date,
  conf_statement
)
}
\arguments{
\item{study}{The name of the study, e.g., "Sponsor Protocol"}

\item{report_title}{The title of the report, e.g., "Open Report"}

\item{purpose}{E.g., "DMC Meeting"}

\item{mtg_type}{What kind of meeting is it? E.g., "Teleconference"}

\item{mtg_date}{The date of the meeting in ISO 8601 E.g., "Friday, December 10, 2021"}

\item{mtg_times}{Character vector of meeting times, E.g., c("7:00 AM - 10:00 AM PST", "10:00 AM - 1:00 PM EST")}

\item{cutoff_snapshot}{"Cutoff" if cutoff date is used "Snapshot" if snapshot date is used}

\item{dco_date}{Data cutoff or snapshot date E.g., "November 10, 2021"}
}
\description{
Generate cover page LaTeX string
}
