% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tex.R
\name{paste_middle}
\alias{paste_middle}
\title{Build LaTeX string of the items to add}
\usage{
paste_middle(tlfs)
}
\arguments{
\item{tlfs}{Data frame with columns for section titles for each TLF
("section", "subsection", "subsubsection") and "path" with the path to the
pdf file for the TLF.}
}
\description{
Build LaTeX string of the items to add
}
