% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tex.R
\name{paste_tex}
\alias{paste_tex}
\title{Build LaTeX string of concatenated report}
\usage{
paste_tex(tlfs, toc, coverpage)
}
\arguments{
\item{tlfs}{Data frame with two columns, "title" with the title of each TLF and "path" with the path to the pdf file for the TLF.}

\item{toc}{Should a table of contents be included?}

\item{coverpage}{A character string of LaTeX to include for a coverpage.}
}
\description{
Build LaTeX string of concatenated report
}
