% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pps.R
\name{read_format_pps}
\alias{read_format_pps}
\title{Read and format the PPS}
\usage{
read_format_pps(path_pps, open_closed)
}
\arguments{
\item{open_closed}{String that identifies which report to produced. Usually
"open" or "closed"}

\item{path_xlsx}{Path to the PPS}
}
\description{
Read and format the PPS
}
