% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_format_tlf_csv.R
\name{read_format_tlf_csv}
\alias{read_format_tlf_csv}
\title{Read and format TLF CSV}
\usage{
read_format_tlf_csv(path_csv)
}
\arguments{
\item{path_csv}{Path to a CSV file with with two columns: "title" with
the title of each TLF and "path" with the path to the pdf file for the TLF.}
}
\description{
Read and format TLF CSV
}
