% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_times.R
\name{remove_leading_zero}
\alias{remove_leading_zero}
\title{Remove leading zero}
\usage{
remove_leading_zero(string)
}
\arguments{
\item{string}{String to remove leading zeros from}
}
\description{
Removes the first zero in a string
}
