% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_times.R
\name{remove_leading_zeros}
\alias{remove_leading_zeros}
\title{Remove leading zeros}
\usage{
remove_leading_zeros(string)
}
\arguments{
\item{string}{String to remove leading zeros from}
}
\description{
Remove leading zeros
}
